requirements.txt
python-interface==1.5.0
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
tensorflow==1.12
requests==2.18.4
scikit-learn==0.19.1
numpy==1.14.3
Flask==1.0.2
Flask-Cors==3.0.4
gunicorn==19.9.0
malaya==1.2.5
google-api-python-client==1.7.7